Moved some stuff into utils file
